Inline getParentFieldElements
